Launcher3: Avoid closing PopUpContainer immediately after view is clicked.
When an item is clicked, we start an app launch animation
from the icon in the container to the full app. The issue is that
we close the container too early, which causes a janky experience.

Tested that we already handle closing the container in all click
instances (ie. when launcher is in background).

Bug: 185510490
Test: click on all the different view types within PopupContainer
      ensure that the container does get closed via logs

Change-Id: I4fef24b911137d0abd24746fa49f466f5183176c
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>